chore(tests): move template helper methods to test-case
