+++ tests/TestCase.php
@@ -4,8 +4,9 @@ declare(strict_types=1);
 
 namespace App\Tests;
 
-use App\Tests\Concern\TemplateHelper;
 use Doctrine\ODM\MongoDB\Types\Type;
+use ReflectionClass;
+use ReflectionMethod;
 
 /**
  * Class TestCase
@@ -15,8 +16,6 @@ use Doctrine\ODM\MongoDB\Types\Type;
  */
 class TestCase extends \PHPUnit\Framework\TestCase
 {
-    use TemplateHelper;
-
     protected function setUp(): void
     {
         parent::setUp();
@@ -45,6 +44,52 @@ class TestCase extends \PHPUnit\Framework\TestCase
         $type::registerType($type, $type);
 
         return $type::getType($type);
+    }
+
+    protected function setUpTraits(): void
+    {
+        $traits = array_keys($this->classUsesRecursive(static::class));
+
+        foreach ($traits as $trait) {
+            $reflection = new ReflectionClass($trait);
+
+            if (! str_starts_with($reflection->getShortName(), 'With')) {
+                continue;
+            }
+
+            $methods = $reflection->getMethods(ReflectionMethod::IS_PROTECTED);
+            foreach ($methods as $method) {
+                if ('setUp' === $method->name) {
+                    continue;
+                }
+                if (str_starts_with($method->name, 'setUp')) {
+                    $this->{$method->name}();
+                }
+            }
+        }
+    }
+
+    protected function tearDownTraits(): void
+    {
+        $traits = array_keys($this->classUsesRecursive(static::class));
+
+        foreach ($traits as $trait) {
+            $reflection = new ReflectionClass($trait);
+
+            if (! str_starts_with($reflection->getShortName(), 'Creates')) {
+                continue;
+            }
+
+            $methods = $reflection->getMethods(ReflectionMethod::IS_PROTECTED);
+            foreach ($methods as $method) {
+                if ('tearDown' === $method->name) {
+                    continue;
+                }
+                if (str_starts_with($method->name, 'tearDown')) {
+                    $this->{$method->name}();
+                }
+            }
+        }
     }
 
     protected function classUsesRecursive(string $class): array
